+++ src/Lexer.h
@@ -87,11 +87,11 @@ typedef struct {
     TkInfo info; // Information about the most recently lexed token
 
     Tk tk; // Most recently lexed token
-    char *ident; int len; // For TK_IDENT
-    uint64_t kint;        // For TK_KINT
-    double kfloat;        // For TK_KFLOAT
-    char kch;             // For TK_KCHAR
-    char *kstr;           // For TK_KSTR
+    char *ident; size_t len; // For TK_IDENT
+    uint64_t kint;           // For TK_KINT
+    double kfloat;           // For TK_KFLOAT
+    char kch;                // For TK_KCHAR
+    char *kstr;              // For TK_KSTR
 } Lexer;
 
 Lexer new_lexer(char *file);
